backend/src/main/resources/templates/backend/django/requirements.txt
Django==5.1.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
django-prometheus==2.3.1
